+++ src/errors.js
@@ -1,8 +1,7 @@
 export class MvxError extends Error {
-  constructor(message, { code = 'MVX_ERROR', cause, details } = {}) {
+  constructor(message, { code = 'MVX_ERROR', cause } = {}) {
     super(message, { cause });
     this.name = 'MvxError';
     this.code = code;
-    if (details !== undefined) this.details = details;
   }
 }